fix(obtenant): modify iops fields of obtenant to int64 type

+++ api/v1alpha1/obtenant_types.go
@@ -85,8 +85,8 @@ type UnitConfig struct {
 	MaxCPU      resource.Quantity `json:"maxCPU"`
 	MemorySize  resource.Quantity `json:"memorySize"`
 	MinCPU      resource.Quantity `json:"minCPU,omitempty"`
-	MaxIops     int               `json:"maxIops,omitempty"`
-	MinIops     int               `json:"minIops,omitempty"`
+	MaxIops     int64             `json:"maxIops,omitempty"`
+	MinIops     int64             `json:"minIops,omitempty"`
 	IopsWeight  int               `json:"iopsWeight,omitempty"`
 	LogDiskSize resource.Quantity `json:"logDiskSize,omitempty"`
 }
